+++ tests/ModelB.php
@@ -2,9 +2,6 @@
 
 namespace Arrayy\tests;
 
-/**
- * Class ModelB
- */
-class ModelB extends \Arrayy\Arrayy
+class ModelB extends \Arrayy\Arrayy implements ModelInterface
 {
 }